Rework verify_phone command
